requirements.txt
torch>=2.0.0
torchvision>=0.15.0
flask>=2.0.0
pillow>=9.0.0
gunicorn>=20.1.0
